// src/app/modules/admin/models/inventory.ts
export class inventory {
  productId: any;
  productName: any;
  categoriesName: any;
  openingStock: any;
  incomingStock: any;
  outgoingStock: any;
  closingStock: any;
  categories_type_id: any;
  sub_categories_id: any;
  getProductId: any;
  setProductId: any;
  stock_id: any;
  warehouse_id: any;
  user_id: any;
  user_role_id: any;
  categories_id: any;
  product_id: any;
  stock_in: any;
  stock_out: any;
  shelve_no: any;
  created_date: any;
  modified_date: any;
  packed_date: any;
  expiry_date: any;
  stock_value: any;
  selling_price: any;
  purchase_price: any;
  my_discount: any;
  available_quantity: any;
  quality: any;
  batch_no: any;
  doc_certified: any;
  stockId: any;
  userId: any;
  sellingPrice: any;
  purchasePrice: any;
  myDiscount: any;
  createdDate: any;
  packedDate: any;
}